Create Electricity Bill.py
Write a program to calculate and print the Electricity bill where the unit consumed by the user is given from test case. It prints the total amount the customer has to pay. The charge are as follows: 
Unit                                                     Charge / Unit
Upto 199                                             @1.20
200 and above but less than 400        @1.50
400 and above but less than 600        @1.80
600 and above                                    @2.00
If bill exceeds Rs.400 then a surcharge of 15% will be charged and the minimum bill should be of Rs.100/-